pyproject.toml
[tool.poetry]
name = "ViPR"
version = "0.1.0"
description = "Virtual PR Review Tool"
authors = ["Katrin Leinweber <kalei@posteo.de>"]

[tool.poetry.dependencies]
python = "^3.7"
pygithub = "^1.45"
click = "^7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
